pyproject.toml
[tool.poetry]
name = "tg_llama2_whisper"
version = "0.1.0"
description = ""
authors = ["Криволапоtrover97"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.260"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
pydub = "^0.25.1"
python-magic = "0.4.27"
telethon = "^1.29.3"
cryptg = "^0.4.0"
setuptools = "^68.2.0"
pillow = "^10.0.0"
faster-whisper = "^0.9.0"
llama-cpp-python = "^0.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
